requirements.txt
langchain>=0.3.0
langchain-core>=1.3.0
langchain-community>=0.4.0
langchain-chroma>=1.1.0
chromadb>=1.5.0
fastembed>=0.8.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0

pymupdf>=1.25.0
pypdf>=5.0.0
pytesseract>=0.3.10
pillow>=10.0.0

python-dotenv>=1.0.0
pydantic>=2.10.0
typing-extensions>=4.12.0

flask>=3.0.0
openai>=1.50.0

fpdf2>=2.7.0